Refactor AIApplyDatatable and related components for improved structure and readability
- Moved column definitions to a separate AIApplyColumn file for better organization.
- Updated AIApplyDatatable to utilize the new column definitions.
- Simplified data fetching and state management in AIJobCustomization and Home components.
- Removed unused imports and commented-out code across various files for cleaner codebase.
- Enhanced user experience by adjusting layout and adding skeleton loaders.
- Implemented a new useDashboardPrefetch hook for efficient data prefetching in the dashboard.
- Removed deprecated proxy file and adjusted related configurations.
