+++ src/app/dashboard/(dashboard)/components/RecentActivityCard.tsx
@@ -22,10 +22,11 @@ import {
 } from "@/components/ui/dropdown-menu";
 import { JobType } from "@/types";
 import { userQueries } from "@/lib/queries/user.queries";
+import { Skeleton } from "@/components/ui/skeleton";
 
 export const RecentActivityCard = memo(
   ({ filters }: { filters: JobFilters }) => {
-    const { data: jobs } = useQuery(jobsQueries.autoApply(filters));
+    const { data: jobs, isLoading } = useQuery(jobsQueries.autoApply(filters));
     const { data: user } = useQuery(userQueries.detail());
     const sortedJobs = useMemo(() => {
       if (!jobs?.data) return [];
@@ -86,101 +87,108 @@ export const RecentActivityCard = memo(
         `/dashboard/tailor-cover-letter/${uuidv4()}?${params}&aiApply=true`
       );
     };
+
     return (
       <Card className="p-4 sm:p-7 gap-4">
         <h1 className="font-bold text-xl">Personalized Recommendation</h1>
         <div className="grid grid-cols-1 lg:grid-cols-2 gap-y-4 sm:gap-8">
-          {sortedJobs?.map((job) => (
-            <div
-              key={job.id}
-              className="flex bg-slate-50 p-4 sm:p-6 rounded-xl gap-4 sm:gap-6 border border-[#cbd5e1] relative"
-            >
-              {job?.emailApply && (
-                <Sparkles className="absolute bottom-4 left-4 text-gray-200 w-5 h-5" />
-              )}
-              <div className="shrink-0">
-                <img
-                  src={!!job.companyLogo ? job.companyLogo : "/placeholder.jpg"}
-                  alt=""
-                  loading="lazy"
-                  className="size-12"
-                />
-              </div>
-              <div className="flex flex-col gap-2 ">
-                <h1 className="font-inter line-clamp-1 capitalize">
-                  {job.title}
-                </h1>
+          {isLoading
+            ? Array.from({ length: 6 }).map((_, index) => (
+                <LoadingSkeleton key={index} />
+              ))
+            : sortedJobs?.map((job) => (
+                <div
+                  key={job.id}
+                  className="flex bg-slate-50 p-4 sm:p-6 rounded-xl gap-4 sm:gap-6 border border-[#cbd5e1] relative"
+                >
+                  {job?.emailApply && (
+                    <Sparkles className="absolute bottom-4 left-4 text-gray-200 w-5 h-5" />
+                  )}
+                  <div className="shrink-0">
+                    <img
+                      src={
+                        !!job.companyLogo ? job.companyLogo : "/placeholder.jpg"
+                      }
+                      alt=""
+                      loading="lazy"
+                      className="size-12"
+                    />
+                  </div>
+                  <div className="flex flex-col gap-2 ">
+                    <h1 className="font-inter line-clamp-1 capitalize">
+                      {job.title}
+                    </h1>
 
-                <DropdownMenu>
-                  <DropdownMenuTrigger
-                    asChild
-                    className="absolute top-4 right-4"
-                  >
-                    <Button variant="ghost">
-                      <MoreHorizontal className="w-4 h-4 text-" />
-                    </Button>
-                  </DropdownMenuTrigger>
-                  <DropdownMenuContent className="w-56" align="start">
-                    <DropdownMenuItem onClick={() => handleJobClick(job)}>
-                      <img src="/cube.svg" className="size-4" alt="" />
-                      Auto apply
-                    </DropdownMenuItem>
-                    <DropdownMenuSeparator />
-                    <DropdownMenuItem
-                      onClick={() => {
-                        router.push(
-                          `/dashboard/jobs/${job.id}?referrer=dashboard&title=${job.title}`
-                        );
-                      }}
-                    >
-                      <img src="/preview.svg" className="size-4" alt="" />
-                      Preview
-                    </DropdownMenuItem>
-                  </DropdownMenuContent>
-                </DropdownMenu>
-                <p className="font-poppins text-cverai-brown text-xs">
-                  <span className="max-w-sm overflow-hidden capitalize">
-                    {job.companyName}
-                  </span>{" "}
-                  ·{" "}
-                  <span className="font-inter text-gray-400">
-                    {!isEmpty(job.salaryInfo) ? job?.salaryInfo : ""}
-                  </span>
-                </p>
-                <div className="flex flex-wrap gap-2">
-                  <div className="flex flex-wrap items-center gap-2">
-                    <Badge
-                      className={cn(
-                        "rounded-full font-epilogue font-semibold bg-cverai-teal/10 text-cverai-teal",
-                        " truncate overflow-hidden text-start "
-                      )}
-                    >
-                      {!!job.jobType ? job.jobType : job?.employmentType}
-                    </Badge>
-                    <div className="bg-slate-500 w-[1px] h-7" />
-                    <Badge
-                      className={cn(
-                        "rounded-full font-epilogue text-wrap font-semibold text-cverai-blue border-cverai-blue bg-white",
-                        " truncate overflow-hidden text-start max-sm:max-w-44"
-                      )}
-                    >
-                      {job.location}
-                    </Badge>
-                    {job?.relevanceScore > 40 && (
-                      <Badge
-                        className={cn(
-                          "rounded-full font-epilogue font-semibold text-cverai-orange border-cverai-orange bg-white",
-                          " truncate overflow-hidden text-start"
-                        )}
+                    <DropdownMenu>
+                      <DropdownMenuTrigger
+                        asChild
+                        className="absolute top-4 right-4"
                       >
-                        {job?.relevanceScore}% match
-                      </Badge>
-                    )}
+                        <Button variant="ghost">
+                          <MoreHorizontal className="w-4 h-4 text-" />
+                        </Button>
+                      </DropdownMenuTrigger>
+                      <DropdownMenuContent className="w-56" align="start">
+                        <DropdownMenuItem onClick={() => handleJobClick(job)}>
+                          <img src="/cube.svg" className="size-4" alt="" />
+                          Auto apply
+                        </DropdownMenuItem>
+                        <DropdownMenuSeparator />
+                        <DropdownMenuItem
+                          onClick={() => {
+                            router.push(
+                              `/dashboard/jobs/${job.id}?referrer=dashboard&title=${job.title}`
+                            );
+                          }}
+                        >
+                          <img src="/preview.svg" className="size-4" alt="" />
+                          Preview
+                        </DropdownMenuItem>
+                      </DropdownMenuContent>
+                    </DropdownMenu>
+                    <p className="font-poppins text-cverai-brown text-xs">
+                      <span className="max-w-sm overflow-hidden capitalize">
+                        {job.companyName}
+                      </span>{" "}
+                      ·{" "}
+                      <span className="font-inter text-gray-400">
+                        {!isEmpty(job.salaryInfo) ? job?.salaryInfo : ""}
+                      </span>
+                    </p>
+                    <div className="flex flex-wrap gap-2">
+                      <div className="flex flex-wrap items-center gap-2">
+                        <Badge
+                          className={cn(
+                            "rounded-full font-epilogue font-semibold bg-cverai-teal/10 text-cverai-teal",
+                            " truncate overflow-hidden text-start "
+                          )}
+                        >
+                          {!!job.jobType ? job.jobType : job?.employmentType}
+                        </Badge>
+                        <div className="bg-slate-500 w-px h-7" />
+                        <Badge
+                          className={cn(
+                            "rounded-full font-epilogue text-wrap font-semibold text-cverai-blue border-cverai-blue bg-white",
+                            " truncate overflow-hidden text-start max-sm:max-w-44"
+                          )}
+                        >
+                          {job.location}
+                        </Badge>
+                        {job?.relevanceScore > 40 && (
+                          <Badge
+                            className={cn(
+                              "rounded-full font-epilogue font-semibold text-cverai-orange border-cverai-orange bg-white",
+                              " truncate overflow-hidden text-start"
+                            )}
+                          >
+                            {job?.relevanceScore}% match
+                          </Badge>
+                        )}
+                      </div>
+                    </div>
                   </div>
                 </div>
-              </div>
-            </div>
-          ))}
+              ))}
         </div>
       </Card>
     );
@@ -188,3 +196,11 @@ export const RecentActivityCard = memo(
 );
 
 RecentActivityCard.displayName = "RecentActivityCard";
+
+const LoadingSkeleton = () => {
+  return (
+    <div className="flex flex-col space-y-3">
+      <Skeleton className="h-34 w-full rounded-xl" />
+    </div>
+  );
+};
