+++ src/app/dashboard/(dashboard)/dashboard-tabs/ai-apply-tab/AIApplyDatatable.tsx
@@ -1,5 +1,4 @@
 import {
-  ColumnDef,
   ColumnFiltersState,
   flexRender,
   getCoreRowModel,
@@ -14,9 +13,7 @@ import {
 } from "@tanstack/react-table";
 import { z } from "zod";
 
-import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
-import { Checkbox } from "@/components/ui/checkbox";
 import {
   Table,
   TableBody,
@@ -26,112 +23,10 @@ import {
   TableRow,
 } from "@/components/ui/table";
 import { Card } from "@/components/ui/card";
-import { cn } from "@/lib/utils";
 import { useRouter } from "next/navigation";
 import { Loader2 } from "lucide-react";
-import { formatFirestoreDate } from "@/lib/utils/helpers";
 import { useMemo, useState } from "react";
-
-export const schema = z.object({
-  id: z.number(),
-  title: z.string(),
-  generatedAt: z.string(),
-  type: z.string(),
-  action: z.string(),
-  recruiterEmail: z.string(),
-  jobDescription: z.string(),
-  coverLetterId: z.string(),
-  resumeId: z.string(),
-});
-
-const getColumns = (
-  router: ReturnType<typeof useRouter>
-): ColumnDef<z.infer<typeof schema>>[] => [
-  {
-    id: "select",
-    header: ({ table }) => (
-      <div className="flex items-center justify-center">
-        <Checkbox
-          checked={
-            table.getIsAllPageRowsSelected() ||
-            (table.getIsSomePageRowsSelected() && "indeterminate")
-          }
-          onCheckedChange={(value) => table.toggleAllPageRowsSelected(!!value)}
-          aria-label="Select all"
-        />
-      </div>
-    ),
-    cell: ({ row }) => (
-      <div className="flex items-center justify-center">
-        <Checkbox
-          checked={row.getIsSelected()}
-          onCheckedChange={(value) => row.toggleSelected(!!value)}
-          aria-label="Select row"
-        />
-      </div>
-    ),
-    enableSorting: false,
-    enableHiding: false,
-  },
-  {
-    accessorKey: "title",
-    header: "Job Title",
-    cell: ({ row }) => {
-      return <div className="font-inter">{row.original.title}</div>;
-    },
-    enableHiding: false,
-  },
-  {
-    accessorKey: "generatedAt",
-    header: "Application Date",
-    cell: ({ row }) => (
-      <div className="">
-        <Badge variant="outline" className={cn("border-0 px-1.5 font-inter")}>
-          {formatFirestoreDate(row.original?.generatedAt)}
-        </Badge>
-      </div>
-    ),
-  },
-  {
-    accessorKey: "type",
-    header: "Application Method",
-    cell: ({ row }) => (
-      <div className="">
-        <div>
-          <Badge
-            variant="outline"
-            className={cn(
-              "text-muted-foreground px-1.5 rounded-2xl font-jakarta capitalize",
-              row.original.type !== "email"
-                ? "bg-primary/10 border-primary/40 text-primary"
-                : "bg-cverai-green/10 border-cverai-green/40 text-cverai-green"
-            )}
-          >
-            {row.original.type}
-          </Badge>
-        </div>
-      </div>
-    ),
-  },
-  {
-    id: "actions",
-    cell: ({ row }) => {
-      return (
-        <Button
-          variant={"ghost"}
-          className="text-blue-500 font-jakarta"
-          onClick={() => {
-            router.push(
-              `/dashboard/preview?aiApplyId=${row.original.id}&resumeId=${row.original.resumeId}&coverLetterId=${row.original.coverLetterId}&recruiterEmail=${row.original.recruiterEmail}&jobDescription=${row.original.jobDescription}`
-            );
-          }}
-        >
-          Show Details
-        </Button>
-      );
-    },
-  },
-];
+import { AIApplyColumn, schema } from "./AIApplyColumn";
 
 export function AIApplyDatatable({
   data,
@@ -141,18 +36,15 @@ export function AIApplyDatatable({
 }) {
   const router = useRouter();
   const [rowSelection, setRowSelection] = useState({});
-  const [columnVisibility, setColumnVisibility] =
-    useState<VisibilityState>({});
-  const [columnFilters, setColumnFilters] = useState<ColumnFiltersState>(
-    []
-  );
+  const [columnVisibility, setColumnVisibility] = useState<VisibilityState>({});
+  const [columnFilters, setColumnFilters] = useState<ColumnFiltersState>([]);
   const [sorting, setSorting] = useState<SortingState>([]);
   const [pagination, setPagination] = useState({
     pageIndex: 0,
     pageSize: 10,
   });
 
-  const columns = useMemo(() => getColumns(router), [router]); // Generate columns with router
+  const columns = useMemo(() => AIApplyColumn(router), [router]); // Generate columns with router
 
   const table = useReactTable({
     data,
@@ -197,7 +89,7 @@ export function AIApplyDatatable({
     table.getRowModel().rows?.length !== 0 && (
       <Card>
         <div className="">
-          <h1 className="font-bold text-xl px-6">Recent Activity</h1>
+          <h1 className="font-bold text-xl px-6 mb-6">Recent Activity</h1>
           <div className="overflow-hidden grid">
             <Table>
               <TableHeader className="bg-muted sticky top-0 z-10">
