+++ src/app/dashboard/(dashboard)/dashboard-tabs/ai-job-customization-tab/AIJobCustomization.tsx
@@ -14,11 +14,14 @@ import { shuffleArray } from "@/lib/utils/helpers";
 
 export const AIJobCustomization = memo(
   ({ filters }: { filters: JobFilters }) => {
-    const { data: resumes } = useQuery(resumeQueries.all(filters));
-    const { data: coverLetter } = useQuery(coverLetterQueries.all(filters));
-    const { data: interviewQuestion } = useQuery(
-      interviewQuestionQueries.all(filters)
+    const { data: resumes } = useQuery(
+      resumeQueries.all(filters)
     );
+    const { data: coverLetter } = useQuery(
+      coverLetterQueries.all(filters)
+    );
+    const { data: interviewQuestion } =
+      useQuery(interviewQuestionQueries.all(filters));
 
     const data = useMemo(() => {
       const merged = [
@@ -38,7 +41,9 @@ export const AIJobCustomization = memo(
 
         <div className="grid gap-y-16">
           <AIJobCustomizationInput />
-          {isEmpty(data) ? null : <AIJobCustomizationDatatable data={data ?? []} />}
+          {isEmpty(data) ? null : (
+            <AIJobCustomizationDatatable data={data ?? []} />
+          )}
           <RecentActivityCard filters={filters} />
         </div>
       </div>
