+++ src/app/dashboard/settings/page.tsx
@@ -1,9 +1,8 @@
-import Settings from "./Settings";
-import { requireOnboarding } from "@/lib/server-auth";
-import { createServerQueryClient } from "@/lib/query/prefetch";
-import { userQueries } from "@/lib/queries/user.queries";
-import { dehydrate } from "@tanstack/react-query";
-import { HydrationBoundary } from "@/components/hydration-boundary";
+// import Settings from "./Settings";
+// import { createServerQueryClient } from "@/lib/query/prefetch";
+// import { userQueries } from "@/lib/queries/user.queries";
+// import { dehydrate } from "@tanstack/react-query";
+// import { HydrationBoundary } from "@/components/hydration-boundary";
 
 const page = async ({
   searchParams,
@@ -11,20 +10,21 @@ const page = async ({
   searchParams: Promise<{ tab: string }>;
 }) => {
   const { tab } = await searchParams;
-  await requireOnboarding();
-  const queryClient = createServerQueryClient();
+  console.log("Settings page tab:", tab);
+  // const queryClient = createServerQueryClient();
 
-   await queryClient.prefetchQuery(
-    userQueries.detail()
-  );
+  //  await queryClient.prefetchQuery(
+  //   userQueries.detail()
+  // );
 
-  const dehydratedState = dehydrate(queryClient);
+  // const dehydratedState = dehydrate(queryClient);
 
   return (
     <div className="p-4 sm:p-8">
-      <HydrationBoundary state={dehydratedState}>
+        Settings Page
+      {/* <HydrationBoundary state={dehydratedState}>
         <Settings tab={tab} />
-      </HydrationBoundary>
+      </HydrationBoundary> */}
     </div>
   );
 };
